requirements.txt
google-auth==2.29.0
google-auth-oauthlib==1.2.0
google-api-python-client==2.127.0
google-generativeai==0.7.2
weasyprint==62.3
xhtml2pdf==0.2.16
python-dotenv==1.0.1
